+++ components/o-audio/test/tracking.test.js
@@ -238,12 +238,12 @@ describe('Tracking' , () => {
 			})
 		);
 
-		it('doesnt allow unknown attributes', () => {
+		it('does allow unknown attributes', () => {
 			const events = oTracking.start();
 			const stubAudioEl = initAudioElement();
 			initTracking(stubAudioEl, { foo: 'bar' });
 			stubAudioEl.dispatchEvent(new Event('playing'));
-			proclaim.notOk(events[0].foo);
+			proclaim.equal(events[0].foo, 'bar');
 		});
 	});
 });